Ver. 0.03
Added:
-The game now recognizes the hand ranks.
--both royal and numbered pairs, trips, and straights. 

Statement:
The calculator functions enough to produce enough information to fulfill its purpose. If logging was added here it would be able to solve the debate over whether straights are more common or trips. However, I believe this code could be expanded to give much more data than originally envisioned. As well i think this code could be repurposed to be the core of an actual playable version of the game.